Speed up coin selection by pre-selecting only as many UTXOs as we need
This effectively converts the coin selection algorithm into "oldest first" (or
whatever sorting is applied to the UTXOs in-memory), but is not exactly that, as
the selected UTXOs are then passed into the existing knapsack algorithm from
Bitcoin. The speedup comes from that algorithm having many fewer UTXOs to
consider (in the case where the number of UTXOs in the wallet is orders of
magnitude greater than the number required).